+++ examples/stallion.py
@@ -2,12 +2,12 @@ from pathlib import Path
 import pickle
 import warnings
 
+import lightning.pytorch as pl
+from lightning.pytorch.callbacks import EarlyStopping, LearningRateMonitor
+from lightning.pytorch.loggers import TensorBoardLogger
 import numpy as np
 import pandas as pd
 from pandas.core.common import SettingWithCopyWarning
-import pytorch_lightning as pl
-from pytorch_lightning.callbacks import EarlyStopping, LearningRateMonitor
-from pytorch_lightning.loggers import TensorBoardLogger
 import torch
 
 from pytorch_forecasting import GroupNormalizer, TemporalFusionTransformer, TimeSeriesDataSet