+++ pytorch_forecasting/optim.py
@@ -130,7 +130,7 @@ class Ranger(Optimizer):
 
     def __setstate__(self, state: dict) -> None:
         super().__setstate__(state)
-        self.radam_buffer = state["radam_buffer"]
+        self.radam_buffer = state.get("radam_buffer", [[None, None, None] for ind in range(10)])
         self.alpha = state["alpha"]
         self.k = state["k"]
         self.N_sma_threshhold = state["N_sma_threshhold"]